+++ CHANGELOG.md
@@ -14,6 +14,7 @@ for further information about branching and tagging conventions.
 - iOS installation guide
 
 #### Fixed
+- A bug when not able to finish KYC recovery
 - Different validation errors for card and crypto address
 
 ## [1.1.0-rc.0] - 2019-09-30